fix(MobileFilter.vue): missing swap type filter
