+++ locales/en.json
@@ -606,6 +606,7 @@
     "onlyVerifiedIdentities": "Only Verified Identities",
     "sale": "Sale",
     "tradeCollection": "Only Collection {0}",
+    "tradeType": "{0} Type",
     "transfer": "Transfer",
     "verified": "Verified"
   },
